refactor: Move MenuDAO, MenuItemVO, and MenuVO to /domain/menu/domain (#23)
