+++ swumeal/src/main/java/com/swumeal/app/domain/menu/service/FileUploadService.java
@@ -1,12 +1,12 @@
 package com.swumeal.app.domain.menu.service;
 
-import com.swumeal.app.dao.MenuDAO;
+import com.swumeal.app.domain.menu.domain.MenuDAO;
+import com.swumeal.app.domain.menu.domain.MenuItemVO;
+import com.swumeal.app.domain.menu.domain.MenuVO;
 import com.swumeal.app.domain.menu.error.MenuErrorCode;
 import com.swumeal.app.domain.menu.error.exception.MenuApiException;
 import com.swumeal.app.domain.menu.vo.MenuDataVo;
 import com.swumeal.app.domain.model.MealTypeEnum;
-import com.swumeal.app.domain.vo.MenuItemVO;
-import com.swumeal.app.domain.vo.MenuVO;
 import com.swumeal.app.global.util.UploadedFileProcessor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
