+++ swumeal/src/main/java/com/swumeal/app/domain/menu/service/MenuReadService.java
@@ -1,6 +1,6 @@
 package com.swumeal.app.domain.menu.service;
 
-import com.swumeal.app.dao.MenuDAO;
+import com.swumeal.app.domain.menu.domain.MenuDAO;
 import com.swumeal.app.domain.menu.dto.MenuListByDateDto;
 import com.swumeal.app.domain.menu.dto.MenuListByTimeDto;
 import com.swumeal.app.domain.menu.dto.request.DateRequestDto;
