+++ swumeal/src/test/java/com/swumeal/app/domain/menu/service/FileUploadServiceTest.java
@@ -1,6 +1,6 @@
 package com.swumeal.app.domain.menu.service;
 
-import com.swumeal.app.dao.MenuDAO;
+import com.swumeal.app.domain.menu.domain.MenuDAO;
 import com.swumeal.app.domain.menu.dto.MenuListByDateDto;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
